Serious Refactoring
Optimized how ProgramNodes interact with their ProgramTree to keep it updated when they are moved. Added hooks to BaseNode._collect_descendants which ProgramNode overrides to have nodes add/remove themselves from the tree when they or their ancestos are added to or popped from a node in the tree. Fixed _fill_out_program.

Added some checks in ProgramNode to ensure programs cannot be modifies while running.